+++ src/components/access-type-edit-section/index.js
@@ -1,0 +1,1 @@
+export { default } from './access-type-edit-section';
